+++ common/tier4_planning_rviz_plugin/CMakeLists.txt
@@ -18,13 +18,13 @@ include_directories(
 
 ament_auto_add_library(tier4_planning_rviz_plugin SHARED
   # path point
-  include/path/display_base.hpp
-  include/path/display.hpp
+  include/tier4_planning_rviz_plugin/path/display_base.hpp
+  include/tier4_planning_rviz_plugin/path/display.hpp
   src/path/display.cpp
   # footprint
-  include/pose_with_uuid_stamped/display.hpp
+  include/tier4_planning_rviz_plugin/pose_with_uuid_stamped/display.hpp
   src/pose_with_uuid_stamped/display.cpp
-  include/mission_checkpoint/mission_checkpoint.hpp
+  include/tier4_planning_rviz_plugin/mission_checkpoint/mission_checkpoint.hpp
   src/mission_checkpoint/mission_checkpoint.cpp
   src/tools/jsk_overlay_utils.cpp
   src/tools/max_velocity.cpp
