+++ common/tier4_planning_rviz_plugin/include/tier4_planning_rviz_plugin/path/display.hpp
@@ -15,7 +15,7 @@
 #ifndef PATH__DISPLAY_HPP_
 #define PATH__DISPLAY_HPP_
 
-#include <path/display_base.hpp>
+#include "tier4_planning_rviz_plugin/path/display_base.hpp"
 
 #include <autoware_auto_planning_msgs/msg/path.hpp>
 #include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
